refactor(emby2Alist&plex2Alist): 拆分配置文件
