+++ emby2Alist/README.md
@@ -5,6 +5,10 @@ date: 2021/09/06 22:00:00
 
 ### 文章更新记录 
 
+2024/05/21
+
+1.拆分配置文件,根据自身倾向选择 exampleConfig 中的配置并参照注释
+
 2024/05/19
 
 1.提供获取软连接真实路径的配置项,前提条件是此程序或容器必须挂载或具有对应目录的读取权限,否则将跳过处理,不生效,
