+++ emby2Alist/docker/docker-compose.yml
@@ -11,6 +11,7 @@ services:
       # 端口映射规则为,宿主机端口:容器内部端口
       # ports:
       #   - 8091:8091
+      #   - 8095:8095
       # 如果遇到海报全部裂开或日志中 permission denied,以下二选一
       # 还不行的话,自己给宿主机 embyCache 读写权限
       # privileged: true
