+++ emby2Alist/nginx/conf.d/exampleConfig/constant-all.js
@@ -1,4 +1,6 @@
-// 全量配置,媒体库混合,本地文件 + rclone/CD2 挂载的 alist 文件 + strm文件
+// 这个总配置单体文件只是备份,生效需要放置在 conf.d 下,且重命名为 constant.js
+
+// 全量配置,媒体库混合,本地文件 + rclone/CD2 挂载的 alist 文件 + strm文件 + 软链接(路径和文件名不一致)
 // export constant allocation
 
 // 必填项,根据实际情况修改下面的设置
