--- emby2Alist/nginx/conf.d/exampleConfig/constant-strm.js
@@ -1,207 +1,0 @@
-// 只使用 strm 文件配置模板,即标准 strm 内部只有远程链接,不存在/开头的相对路径
-// 不需要挂载功能,不显示依赖 alist,strm 内部为任意直链
-// export constant allocation
-
-// 必填项,根据实际情况修改下面的设置
-
-// 这里默认 emby/jellyfin 的地址是宿主机,要注意 iptables 给容器放行端口
-const embyHost = "http://172.17.0.1:8096";
-
-// emby/jellyfin api key, 在 emby/jellyfin 后台设置
-const embyApiKey = "f839390f50a648fd92108bc11ca6730a";
-
-// 选填项,用不到保持默认即可
-
-// 字符串头,用于特殊匹配判断
-const strHead = {
-  lanIp: ["172.", "10.", "192.", "[fd00:"], // 局域网ip头
-  "115": "115.com",
-  ali: "aliyundrive.net",
-};
-
-// 路由缓存配置
-const routeCacheConfig = {
-  enable: false,
-  // 缓存键表达式,默认值好处是命中范围大,但会导致 routeRule 中针对设备的规则失效,多个变量可自行组合修改,冒号分隔
-  // 注意 jellyfin 是小写开头 mediaSourceId
-  keyExpression: "r.uri:r.args.MediaSourceId", // "r.uri:r.args.MediaSourceId:r.args.X-Emby-Device-Id"
-};
-
-// 指定需要获取符号链接真实路径的规则,优先级在 embyMountPath 和 routeRule 之间
-// 注意前提条件是此程序或容器必须挂载或具有对应目录的读取权限,否则将跳过处理,不生效
-// 参数1: 0: startsWith(str), 1: endsWith(str), 2: includes(str), 3: match(/ain/g)
-// 参数2: 匹配目标,对象为媒体服务入库的文件路径(Item.Path)
-const symlinkRule = [
-  // [0, "/mnt/sda1"],
-];
-
-// 路由规则,注意有先后顺序,"proxy"规则优先级最高,其余依次,千万注意规则不要重叠,不然排错十分困难,字幕和图片走了缓存,不在此规则内
-// 参数1: 指定处理模式,单规则的默认值为"proxy",但是注意整体规则都不匹配默认值为"redirect",然后下面参数序号-1
-// "proxy": 原始媒体服务器处理(中转流量), "redirect": 直链302, "transcode": 转码, "block": 只是屏蔽播放
-// "transcode",稍微有些歧义,这里只是不做转码参数修改,具体是否转码由 emby 客户端自己判断上报或客户端手动切换码率控制
-// 参数2: 分组名,组内为与关系(全部匹配),多个组和没有分组的规则是或关系(任一匹配),然后下面参数序号-1
-// 参数3: 匹配类型或来源(字符串参数类型) "filePath": 文件路径(Item.Path), "alistRes": alist返回的链接
-// 参数4: 0: startsWith(str), 1: endsWith(str), 2: includes(str), 3: match(/ain/g)
-// 参数5: 匹配目标,为数组的多个参数时,数组内为或关系(任一匹配)
-const routeRule = [
-  // ["filePath", 0, "/mnt/sda1"],
-  // ["filePath", 1, ".mp3"],
-  // ["filePath", 2, "Google"],
-  // ["alistRes", 2, "/NAS/"], // 例如使用 alias 聚合了 nas 本地文件,可能会存在卡顿或花屏
-  // ["filePath", 3, /private/ig],
-  // docker 注意必须为 host 模式,不然此变量全部为内网ip,判断无效,nginx 内置变量不带$,客户端地址($remote_addr)
-  // ["r.variables.remote_addr", 0, strHead.lanIp],
-  // ["r.headersIn.User-Agent", 2, "IE"], // 请求头参数,客户端UA
-  // ["r.args.X-Emby-Device-Id", 0, "d4f30461-ec5c-488d-b04a-783e6f419eb1"], // 链接入参,设备id
-  // ["r.args.X-Emby-Device-Name", 0, "Microsoft Edge Windows"], // 链接入参,设备名称
-  // ["r.args.UserId", 0, "ac0d220d548f43bbb73cf9b44b2ddf0e"], // 链接入参,用户id
-  // 以下规则代表禁用["Emby Web", "Emby for iOS", "Infuse"]中的[本地挂载文件或 alist 返回的链接]的 115 直链功能
-  // ["115-alist", "r.args.X-Emby-Client", 0, ["Emby Web", "Emby for iOS", "Infuse"]], // 链接入参,客户端类型
-  // ["115-alist", "alistRes", 0, strHead["115"]],
-  // ["115-local", "r.args.X-Emby-Client", 0, ["Emby Web", "Emby for iOS", "Infuse"]],
-  // ["115-local", "filePath", 0, "/mnt/115"],
-  // 注意非"proxy"无法使用"alistRes"条件,因为没有获取 alist 直链的过程
-  // ["proxy", "filePath", 0, "/mnt/sda1"],
-  // ["redirect", "filePath", 0, "/mnt/sda2"],
-  // ["transcode", "filePath", 0, "/mnt/sda3"],
-  // ["transcode", "115-local", "r.args.X-Emby-Client", 0, ["Emby Web", "Emby for iOS", "Infuse"]],
-  // ["transcode", "115-local", "filePath", 0, "/mnt/115"],
-  // ["block", "filePath", 0, "/mnt/sda4"],
-];
-
-// 路径映射,会在 xxxMountPath 之后从上到下依次全部替换一遍,不要有重叠
-// 参数1: 0: 默认做字符串替换, 1: 前插, 2: 尾插
-// 参数2: 0: 默认只处理/开头的路径且不为 strm, 1: 只处理 strm 内部为/开头的相对路径, 2: 只处理 strm 内部为远程链接的
-// 参数3: 来源, 参数4: 目标
-const embyPathMapping = [
-  // [0, 0, "/mnt/aliyun-01", "/mnt/aliyun-02"],
-  // [0, 2, "http:", "https:"], 
-  // [0, 2, ":5244", "/alist"], 
-  // [0, 0, "D:", "F:"],
-  // [0, 0, /blue/g, "red"], // 此处正则不要加引号
-  // [1, 1, `${alistPublicAddr}/d`],
-  // [2, 2, "?xxx"],
-];
-
-// 指定是否转发由 njs 获取 strm 重定向后直链地址的规则,例如 strm 内部为局域网 ip 或链接需要验证
-// 参数1: 0: startsWith(str), 1: endsWith(str), 2: includes(str), 3: match(/ain/g)
-// 参数2: 匹配目标,对象为 xxxPathMapping 映射后的 strm 内部链接
-const redirectStrmLastLinkRule = [
-  [0, strHead.lanIp.map(s => "http://" + s)],
-  // [0, alistAddr],
-  // [0, "http:"],
-  // 参数3: 请求验证类型,当前 alistAddr 不需要此参数
-  // 参数4: 当前 alistAddr 不需要此参数,alistSignExpireTime
-  // [0, "http://otheralist1.com", "sign", `${alistToken}:${alistSignExpireTime}`],
-];
-
-// 图片缓存策略,包括主页、详情页、图片库的原图,路由器 nginx 请手动调小 conf 中 proxy_cache_path 的 max_size
-// 0: 不同尺寸设备共用一份缓存,先访问先缓存,空间占用最小但存在小屏先缓存大屏看的图片模糊问题
-// 1: 不同尺寸设备分开缓存,空间占用适中,命中率低下,但契合 emby 的图片缩放处理
-// 2: 不同尺寸设备共用一份缓存,空间占用最大,移除 emby 的缩放参数,直接原图高清显示
-const imageCachePolicy = 0;
-
-// 对接 emby 通知管理员设置,目前只发送是否直链成功,依赖 emby/jellyfin 的 webhook 配置并勾选外部通知
-const embyNotificationsAdmin = {
-  enable: false,
-  includeUrl: false, // 链接太长,默认关闭
-  name: "【emby2Alist】",
-};
-
-// 对接 emby 设备控制推送通知消息,目前只发送是否直链成功,此处为统一开关,范围为所有的客户端,通知目标只为当前播放的设备
-const embyRedirectSendMessage = {
-  enable: false,
-  header: "【emby2Alist】",
-  timeoutMs: -1, // 消息通知弹窗持续毫秒值
-};
-
-// 按路径匹配规则隐藏部分接口返回的 items
-// 参数1: 0: startsWith(str), 1: endsWith(str), 2: includes(str), 3: match(/ain/g)
-// 参数2: 匹配目标,对象为 Item.Path
-// 参数3: 0: 默认同时过滤下列所有类型接口, 1: 只隐藏[搜索建议(不会过滤搜索接口)]接口,
-// 2: 只隐藏[更多类似(若当前浏览项目位于规则中,将跳过隐藏)]接口, 3: 只隐藏第三方使用的[海报推荐]接口
-const itemHiddenRule = [
-  // [0, "/mnt/sda1"],
-  // [1, ".mp3", 1],
-  // [2, "Google", 2],
-  // [3, /private/ig],
-];
-
-// 串流配置
-const streamConfig = {
-  // 启用后将修改直接串流链接为真实文件名,方便第三方播放器友好显示和匹配,
-  // 默认不启用,可能存在兼容问题,如发现原始链接代理失败,请关闭此选项
-  useRealFileName: false,
-};
-
-// 留空项,不需要更改
-// 挂载工具 rclone/CD2 多出来的挂载目录, 例如将 od,gd 挂载到 /mnt 目录下: /mnt/onedrive /mnt/gd ,那么这里就填写 /mnt
-// 通常配置一个远程挂载根路径就够了,默认非此路径开头文件将转给原始 emby 处理,不用重复填写至 disableRedirectRule
-// 如果没有挂载,全部使用 strm 文件,此项填[""],必须要是数组
-const embyMountPath = [""];
-// 访问宿主机上 5244 端口的 alist 地址, 要注意 iptables 给容器放行端口
-const alistAddr = "";
-// alist token, 在 alist 后台查看
-const alistToken = "";
-// alist 是否启用了 sign
-const alistSignEnable = false;
-// alist 中设置的直链过期时间,以小时为单位
-const alistSignExpireTime = 12;
-// alist公网地址, 用于需要 alist server 代理流量的情况, 按需填写
-const alistPublicAddr = "";
-// 指定客户端自己请求并获取 alist 直链的规则,代码优先级在 redirectStrmLastLinkRule 之后
-// 特殊情况使用,则此处必须使用域名且公网畅通,用不着请保持默认
-// 参数1: 0: startsWith(str), 1: endsWith(str), 2: includes(str), 3: match(/ain/g)
-// 参数2: 匹配目标,对象为 Alist 接口返回的链接 raw_url
-// 参数3: 指定转发给客户端的 alist 的 host 前缀,兼容 sign 参数
-const cilentSelfAlistRule = [];
-// !!!实验功能,转码配置,默认 false,将按之前逻辑禁止转码处理并移除转码选项参数,与 emby 配置无关
-// 主库和所有从库给用户开启[播放-如有必要，在媒体播放期间允许视频转码]+[倒数7行-允许媒体转换]
-// type: "nginx", nginx 负载均衡,好处是使用简单且内置均衡参数选择,缺点是流量全部经过此服务器,
-// 且使用条件很苛刻,转码服务组中的媒体 id 需要和主媒体库中 id 一致,自行寻找实现主从同步,完全同步后,ApiKey 也是一致的
-// type: "distributed-media-server", 分布式媒体服务负载均衡(暂未实现均衡),优先利用 302 真正实现流量的 LB,且灵活,
-// 不区分主从,当前访问服务即为主库,可 emby/jellyfin 混搭,挂载路径可以不一致,但要求库中的标题和语种一致且原始文件名一致
-const transcodeConfig = {
-  enable: false
-};
-
-// for js_set
-function getEmbyHost(r) {
-  return embyHost;
-}
-function getTranscodeEnable(r) {
-  return transcodeConfig.enable;
-}
-function getTranscodeType(r) {
-  return transcodeConfig.type;
-}
-function getImageCachePolicy(r) {
-  return imageCachePolicy;
-}
-
-export default {
-  embyHost,
-  embyMountPath,
-  embyApiKey,
-  alistAddr,
-  alistToken,
-  alistSignEnable,
-  alistSignExpireTime,
-  alistPublicAddr,
-  strHead,
-  routeCacheConfig,
-  symlinkRule,
-  routeRule,
-  embyPathMapping,
-  redirectStrmLastLinkRule,
-  cilentSelfAlistRule,
-  transcodeConfig,
-  embyNotificationsAdmin,
-  embyRedirectSendMessage,
-  itemHiddenRule,
-  streamConfig,
-  getEmbyHost,
-  getTranscodeEnable,
-  getTranscodeType,
-  getImageCachePolicy,
-}
