+++ plex2Alist/nginx/conf.d/config/constant-ext.js
@@ -1,0 +1,6 @@
+
+// 选填项,特定平台功能,用不到保持默认即可
+
+export default {
+
+}
